♻️ Refactor `PostForm` component to encapsulate all related UI elements
Moved the current user avatar and the container from the `ActivityFeed` component to `PostForm` to better adhere to separation of concerns.
